Don't cache; separate from interpreter; use /sys/module/nvidia/version
