+++ crates/uv-platform-tags/Cargo.toml
@@ -16,7 +16,6 @@ doctest = false
 workspace = true
 
 [dependencies]
-uv-pep440 = { workspace = true }
 uv-small-str = { workspace = true }
 
 memchr = { workspace = true }
