+++ crates/uv-platform-tags/src/lib.rs
@@ -1,12 +1,10 @@
 pub use abi_tag::{AbiTag, ParseAbiTagError};
-pub use accelerator::Accelerator;
 pub use language_tag::{LanguageTag, ParseLanguageTagError};
 pub use platform::{Arch, Os, Platform, PlatformError};
 pub use platform_tag::{ParsePlatformTagError, PlatformTag};
 pub use tags::{BinaryFormat, IncompatibleTag, TagCompatibility, TagPriority, Tags, TagsError};
 
 mod abi_tag;
-mod accelerator;
 mod language_tag;
 mod platform;
 mod platform_tag;
