+++ crates/uv-static/src/env_vars.rs
@@ -678,4 +678,11 @@ impl EnvVars {
     ///
     /// This is a quasi-standard variable, described, e.g., in `ncurses(3x)`.
     pub const COLUMNS: &'static str = "COLUMNS";
+
+    /// The CUDA driver version to assume when inferring the PyTorch backend.
+    #[attr_hidden]
+    pub const UV_CUDA_DRIVER_VERSION: &'static str = "UV_CUDA_DRIVER_VERSION";
+
+    /// Equivalent to the `--torch-backend` command-line argument (e.g., `cpu`, `cu126`, or `auto`).
+    pub const UV_TORCH_BACKEND: &'static str = "UV_TORCH_BACKEND";
 }
