+++ crates/uv-torch/Cargo.toml
@@ -14,11 +14,15 @@ uv-distribution-types = { workspace = true }
 uv-normalize = { workspace = true }
 uv-pep440 = { workspace = true }
 uv-platform-tags = { workspace = true }
+uv-static = { workspace = true }
 
 clap = { workspace = true, optional = true }
 either = { workspace = true }
+fs-err = { workspace = true }
 schemars = { workspace = true, optional = true }
 serde = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
 
 [lints]
 workspace = true
